pyproject.toml
[project]
name = "comfy-models"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.10",
    "click>=8.1.7",
    "modal>=0.65.5",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.3",
    "setuptools>=75.3.0",
    "tabulate>=0.9.0",
]

[options.entry_points]
console_scripts = [
    "models=comfy_models.cli:cli"
]

[project.scripts]
models = "comfy_models.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
